Move task.internal.syslog to its own module
